requirements.txt
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
chardet==3.0.4
cssselect==1.1.0
cssutils==2.2.0
idna==2.8
lxml==4.6.3
notify-run==0.0.13
premailer==3.8.0
prettytable==2.1.0
pycodestyle==2.7.0
PyQRCode==1.2.1
python-dotenv==0.17.1
requests==2.22.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.25.11
wcwidth==0.2.5
yagmail==0.14.256
